+++ scenarioo-server/src/main/java/org/scenarioo/rest/application/ScenariooRestApplication.java
@@ -37,7 +37,6 @@ public class ScenariooRestApplication extends Application {
 
 	public ScenariooRestApplication() {
 		// find . -name \*Resource.java
-		//does the order play a role in what order URLs get matched??
 		LOGGER.info("Register Scenarioo REST Services ...");
 		singletons.add(new BranchAliasesResource());
 		singletons.add(new ConfigurationResource());
@@ -50,6 +49,7 @@ public class ScenariooRestApplication extends Application {
 		singletons.add(new BranchBuildsResource());
 		singletons.add(new BuildsResource());
 		singletons.add(new ComparisonsResource());
+		singletons.add(new ComparisonCalculationsResource());
 		singletons.add(new SearchResource());
 		singletons.add(new ScenariosResource());
 		singletons.add(new ScreenshotResource());